Split item types into multiple modules
