+++ emergence_lib/src/items/errors.rs
@@ -1,0 +1,29 @@
+//! Errors related to items and inventories.
+
+use super::count::ItemCount;
+
+/// Failed to add items to an inventory.
+#[derive(Debug)]
+pub struct AddOneItemError {
+    /// The number of items that exceed the capacity.
+    pub excess_count: ItemCount,
+}
+
+/// Failed to add items to an inventory.
+pub struct AddManyItemsError {
+    /// The number of items that exceeded the capacity.
+    pub excess_counts: Vec<ItemCount>,
+}
+
+/// Failed to remove items from an item slot.
+#[derive(Debug)]
+pub struct RemoveOneItemError {
+    /// The number of items that were missing from the inventory.
+    pub missing_count: usize,
+}
+
+/// Failed to remove many items from an inventory.
+pub struct RemoveManyItemsError {
+    /// The number of items that were missing from the inventory.
+    pub missing_counts: Vec<ItemCount>,
+}
