+++ emergence_lib/src/items/recipe.rs
@@ -1,0 +1,63 @@
+//! Instructions to craft items.
+
+use std::{fmt::Display, time::Duration};
+
+use super::count::ItemCount;
+
+/// A recipe to turn a set of items into different items.
+#[derive(Debug, Clone)]
+pub struct Recipe {
+    /// The inputs needed to craft the recipe.
+    inputs: Vec<ItemCount>,
+
+    /// The outputs generated by crafting.
+    outputs: Vec<ItemCount>,
+
+    /// The time needed to craft the recipe.
+    craft_time: Duration,
+}
+
+impl Recipe {
+    /// Create a new recipe with the given inputs, outputs and craft time.
+    pub fn new(inputs: Vec<ItemCount>, outputs: Vec<ItemCount>, craft_time: Duration) -> Self {
+        Self {
+            inputs,
+            outputs,
+            craft_time,
+        }
+    }
+
+    /// The inputs needed to craft the recipe.
+    pub fn inputs(&self) -> &Vec<ItemCount> {
+        &self.inputs
+    }
+
+    /// The outputs generated by crafting.
+    pub fn outputs(&self) -> &Vec<ItemCount> {
+        &self.outputs
+    }
+
+    /// The time needed to craft the recipe.
+    pub fn craft_time(&self) -> &Duration {
+        &self.craft_time
+    }
+}
+
+impl Display for Recipe {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let input_strings: Vec<String> =
+            self.inputs.iter().map(|input| format!("{input}")).collect();
+        let input_str = input_strings.join(", ");
+
+        let output_strings: Vec<String> = self
+            .outputs
+            .iter()
+            .map(|output| format!("{output}"))
+            .collect();
+        let output_str = output_strings.join(", ");
+
+        let duration_str = format!("{:.2}", self.craft_time().as_secs_f32());
+
+        write!(f, "[{input_str}] -> [{output_str}] | {duration_str}s")
+    }
+}
