+++ emergence_lib/src/organisms/sessile/plants.rs
@@ -2,7 +2,7 @@
 
 use crate::{
     self as emergence_lib,
-    items::{ItemCount, ItemId},
+    items::{count::ItemCount, ItemId},
     organisms::life_cycles::LifeCycle,
 };
 use bevy::prelude::*;
@@ -12,7 +12,7 @@ use emergence_macros::IterableEnum;
 use crate::{
     enum_iter::IterableEnum,
     graphics::{organisms::OrganismSprite, sprites::IntoSprite, Tilemap},
-    items::Recipe,
+    items::recipe::Recipe,
     organisms::Species,
 };
